\name{TOMO.colors}
\alias{TOMO.colors}

\title{Tomography Colors}
\description{
 Color Palette ranging from red to blue through black.
}
\usage{
TOMO.colors(n, alpha = 1)
}

\arguments{
  \item{n}{number of colors }
  \item{alpha}{hsv color parameter}
}

\value{
color palette
}

\author{Jonathan M. Lees<jonathan.lees.edu>}

\seealso{rainbow, colors, hsv}
\examples{
TOMO.colors(25, alpha = 1)
}

\keyword{misc}
